fetch avatars clientside
